* [#1758] doc(time): add some comment in the class description * [#1758] refactor(time): split CronExpression in its own file
