+++ framework/src/play/jobs/JobsPlugin.java
@@ -5,18 +5,23 @@ import java.io.StringWriter;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 import java.util.LinkedList;
-import java.util.concurrent.*;
+import java.util.List;
+import java.util.concurrent.Callable;
+import java.util.concurrent.FutureTask;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 import play.Logger;
 import play.Play;
 import play.PlayPlugin;
 import play.exceptions.PlayException;
 import play.exceptions.UnexpectedException;
+import play.libs.CronExpression;
 import play.libs.Expression;
 import play.libs.Time;
-import play.libs.Time.CronExpression;
 import play.mvc.Http.Request;
 import play.utils.Java;
 import play.utils.PThreadFactory;
@@ -46,16 +51,17 @@ public class JobsPlugin extends PlayPlugin {
         SimpleDateFormat df = new SimpleDateFormat("MM/dd/yyyy HH:mm:ss");
         if (!scheduledJobs.isEmpty()) {
             out.println();
-            out.println("Scheduled jobs ("+scheduledJobs.size()+"):");
+            out.println("Scheduled jobs (" + scheduledJobs.size() + "):");
             out.println("~~~~~~~~~~~~~~~~~~~~~~~~~~");
             for (Job job : scheduledJobs) {
                 out.print(job.getClass().getName());
-                if (job.getClass().isAnnotationPresent(OnApplicationStart.class) && !(job.getClass().isAnnotationPresent(On.class) || job.getClass().isAnnotationPresent(Every.class))) {
+                if (job.getClass().isAnnotationPresent(OnApplicationStart.class)
+                        && !(job.getClass().isAnnotationPresent(On.class) || job.getClass().isAnnotationPresent(Every.class))) {
                     OnApplicationStart appStartAnnotation = job.getClass().getAnnotation(OnApplicationStart.class);
-                    out.print(" run at application start" + (appStartAnnotation.async()?" (async)" : "") + ".");
+                    out.print(" run at application start" + (appStartAnnotation.async() ? " (async)" : "") + ".");
                 }
 
-                if( job.getClass().isAnnotationPresent(On.class)) {
+                if (job.getClass().isAnnotationPresent(On.class)) {
 
                     String cron = job.getClass().getAnnotation(On.class).value();
                     if (cron != null && cron.startsWith("cron.")) {
@@ -68,7 +74,7 @@ public class JobsPlugin extends PlayPlugin {
                 }
                 if (job.lastRun > 0) {
                     out.print(" (last run at " + df.format(new Date(job.lastRun)));
-                    if(job.wasError) {
+                    if (job.wasError) {
                         out.print(" with error)");
                     } else {
                         out.print(")");
@@ -87,12 +93,12 @@ public class JobsPlugin extends PlayPlugin {
 
             for (int i = 0; i < q.length; i++) {
                 ScheduledFuture task = q[i];
-                out.println(Java.extractUnderlyingCallable((FutureTask<?>) task) + " will run in " + task.getDelay(TimeUnit.SECONDS) + " seconds");
+                out.println(Java.extractUnderlyingCallable((FutureTask<?>) task) + " will run in " + task.getDelay(TimeUnit.SECONDS)
+                        + " seconds");
             }
         }
         return sw.toString();
     }
-
 
     @Override
     public void afterApplicationStart() {
@@ -106,16 +112,16 @@ public class JobsPlugin extends PlayPlugin {
         for (final Class<?> clazz : jobs) {
             // @OnApplicationStart
             if (clazz.isAnnotationPresent(OnApplicationStart.class)) {
-                //check if we're going to run the job sync or async
+                // check if we're going to run the job sync or async
                 OnApplicationStart appStartAnnotation = clazz.getAnnotation(OnApplicationStart.class);
-                if( !appStartAnnotation.async()) {
-                    //run job sync
+                if (!appStartAnnotation.async()) {
+                    // run job sync
                     try {
                         Job<?> job = ((Job<?>) clazz.newInstance());
                         scheduledJobs.add(job);
                         job.run();
-                        if(job.wasError) {
-                            if(job.lastException != null) {
+                        if (job.wasError) {
+                            if (job.lastException != null) {
                                 throw job.lastException;
                             }
                             throw new RuntimeException("@OnApplicationStart Job has failed");
@@ -131,13 +137,13 @@ public class JobsPlugin extends PlayPlugin {
                         throw new UnexpectedException(ex);
                     }
                 } else {
-                    //run job async
+                    // run job async
                     try {
                         Job<?> job = ((Job<?>) clazz.newInstance());
                         scheduledJobs.add(job);
-                        //start running job now in the background
+                        // start running job now in the background
                         @SuppressWarnings("unchecked")
-                        Callable<Job> callable = (Callable<Job>)job;
+                        Callable<Job> callable = (Callable<Job>) job;
                         executor.submit(callable);
                     } catch (InstantiationException ex) {
                         throw new UnexpectedException("Cannot instanciate Job " + clazz.getName());
@@ -169,7 +175,7 @@ public class JobsPlugin extends PlayPlugin {
                         value = Play.configuration.getProperty(value);
                     }
                     value = Expression.evaluate(value, value).toString();
-                    if(!"never".equalsIgnoreCase(value)){
+                    if (!"never".equalsIgnoreCase(value)) {
                         executor.scheduleWithFixedDelay(job, Time.parseDuration(value), Time.parseDuration(value), TimeUnit.SECONDS);
                     }
                 } catch (InstantiationException ex) {
@@ -206,17 +212,19 @@ public class JobsPlugin extends PlayPlugin {
             CronExpression cronExp = new CronExpression(cron);
             Date nextDate = cronExp.getNextValidTimeAfter(now);
             if (nextDate == null) {
-                Logger.warn("The cron expression for job %s doesn't have any match in the future, will never be executed", job.getClass().getName());
+                Logger.warn("The cron expression for job %s doesn't have any match in the future, will never be executed", job.getClass()
+                        .getName());
                 return;
             }
             if (nextDate.equals(job.nextPlannedExecution)) {
                 // Bug #13: avoid running the job twice for the same time
-                // (happens when we end up running the job a few minutes before the planned time)
+                // (happens when we end up running the job a few minutes before
+                // the planned time)
                 Date nextInvalid = cronExp.getNextInvalidTimeAfter(nextDate);
                 nextDate = cronExp.getNextValidTimeAfter(nextInvalid);
             }
             job.nextPlannedExecution = nextDate;
-            executor.schedule((Callable<V>)job, nextDate.getTime() - now.getTime(), TimeUnit.MILLISECONDS);
+            executor.schedule((Callable<V>) job, nextDate.getTime() - now.getTime(), TimeUnit.MILLISECONDS);
             job.executor = executor;
         } catch (Exception ex) {
             throw new UnexpectedException(ex);
@@ -260,23 +268,23 @@ public class JobsPlugin extends PlayPlugin {
 
     @Override
     public void beforeInvocation() {
-      afterInvocationActions.set(new LinkedList<Callable<? extends Object>>());
+        afterInvocationActions.set(new LinkedList<Callable<? extends Object>>());
     }
 
     @Override
     public void afterInvocation() {
-      List<Callable<? extends Object>> currentActions = afterInvocationActions.get();
-      afterInvocationActions.set(null);
-      for (Callable<? extends Object> callable : currentActions) {
-        JobsPlugin.executor.submit(callable);
-      }
+        List<Callable<? extends Object>> currentActions = afterInvocationActions.get();
+        afterInvocationActions.set(null);
+        for (Callable<? extends Object> callable : currentActions) {
+            JobsPlugin.executor.submit(callable);
+        }
     }
 
     // default visibility, because we want to use this only from Job.java
     static void addAfterRequestAction(Callable<? extends Object> c) {
-      if (Request.current() == null) {
-        throw new IllegalStateException("After request actions can be added only from threads that serve requests!");
-      }
-      afterInvocationActions.get().add(c);
+        if (Request.current() == null) {
+            throw new IllegalStateException("After request actions can be added only from threads that serve requests!");
+        }
+        afterInvocationActions.get().add(c);
     }
 }
